- Deleting Internal Navigation object and using constant instead - Deleting mapper file and adding a constructor in PokemonDB - Deleting setOnClick to receive it through constructor to avoid errors like using it without initialize it - Extracting State classes from ViewModels to independent files
