+++ core/src/main/java/com/bsarias/pokeapi/core/framework/navigation/InternalNavigation.kt
@@ -1,7 +1,5 @@
 package com.bsarias.pokeapi.core.framework.navigation
 
-object InternalNavigation {
-    const val DOMAIN = "myapp://"
-    const val LISTING = "${DOMAIN}listpokemon"
-    const val DETAILS = "${DOMAIN}details?pokemonName="
-}
+const val DOMAIN = "myapp://"
+const val LISTING = "${DOMAIN}listpokemon"
+const val DETAILS = "${DOMAIN}details?pokemonName="
