+++ details/src/main/java/com/bsarias/pokeapi/details/framework/persistence/DetailsPersistenceDataSourceImpl.kt
@@ -2,7 +2,7 @@ package com.bsarias.pokeapi.details.framework.persistence
 
 import com.bsarias.pokeapi.core.domain.Pokemon
 import com.bsarias.pokeapi.core.framework.persistence.daos.PokemonDao
-import com.bsarias.pokeapi.core.framework.persistence.mappers.mapToEntity
+import com.bsarias.pokeapi.core.framework.persistence.entities.PokemonDB
 import com.bsarias.pokeapi.details.data.source.DetailsPersistenceDataSource
 
 class DetailsPersistenceDataSourceImpl(private val dao: PokemonDao) : DetailsPersistenceDataSource {
@@ -15,6 +15,6 @@ class DetailsPersistenceDataSourceImpl(private val dao: PokemonDao) : DetailsPer
     }
 
     override suspend fun savePokemon(pokemon: Pokemon) {
-        dao.savePokemon(pokemon.mapToEntity())
+        dao.savePokemon(PokemonDB(pokemon))
     }
 }
