+++ listpokemon/src/main/java/com/bsarias/pokeapi/list/presentation/ListViewModel.kt
@@ -4,7 +4,6 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.bsarias.pokeapi.core.domain.ListPokemon
 import com.bsarias.pokeapi.list.usecases.GetListPokemons
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.catch
@@ -16,11 +15,6 @@ import javax.inject.Inject
 class ListViewModel @Inject constructor(private val getListPokemons: GetListPokemons) :
     ViewModel() {
 
-    sealed class ListViewState {
-        object Loading : ListViewState()
-        class Success(val listPokemon: ListPokemon) : ListViewState()
-        class Error(val error: String) : ListViewState()
-    }
 
     private val model: MutableLiveData<ListViewState> by lazy {
         MutableLiveData<ListViewState>().also {
