+++ listpokemon/src/main/java/com/bsarias/pokeapi/list/presentation/ListViewState.kt
@@ -1,0 +1,9 @@
+package com.bsarias.pokeapi.list.presentation
+
+import com.bsarias.pokeapi.core.domain.ListPokemon
+
+sealed class ListViewState {
+    object Loading : ListViewState()
+    class Success(val listPokemon: ListPokemon) : ListViewState()
+    class Error(val error: String) : ListViewState()
+}
